refactor: replace with SearchController.serachRecipes
