+++ backend/src/app.ts
@@ -2,6 +2,7 @@ import express from "express"
 import cors from "cors"
 import { client } from "./db.server"
 import UserController from "./controllers/UserController"
+import SearchController from "./controllers/SearchController"
 
 const app = express()
 
@@ -14,55 +15,7 @@ app.use(cors())
 app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
 
-export type SearchInfo = {
-  ingredients: string[]
-  dish: string | null
-  cookingTime: string | null
-}
-
-app.post("/api/searchRecipes", async (req, res) => {
-  const { searchInfo } = req.body
-  const ingredientsAndQuery: string = searchInfo.ingredients.join(" & ")
-  const dishQuery = searchInfo.dish
-  let cookingTimeQuery = {}
-  switch (searchInfo.cookingTime) {
-    case "時短":
-      cookingTimeQuery = { gte: 1, lt: 20 }
-      break
-    case "普通":
-      cookingTimeQuery = { gte: 20, lt: 45 }
-      break
-    case "じっくり":
-      cookingTimeQuery = { gte: 45 }
-      break
-    default:
-      break
-  }
-
-  const orderBy = Math.random() < 0.5 ? "description" : "materialsConverted"
-  const orderDir = Math.random() < 0.5 ? "asc" : "desc"
-  // const itemCount = await client.recipes.count()
-  // const skip = Math.max(0, Math.floor(Math.random() * itemCount) - 5000)
-
-  const recipes = await client.recipes.findMany({
-    orderBy: {
-      [orderBy]: orderDir,
-    },
-    // FIXME: skip をすると全部消えてしまう, ランダムサンプリングの書き方が弱い(rawSQLにする?)
-    // skip: skip,
-    where: {
-      materialsConverted: {
-        search: ingredientsAndQuery,
-      },
-      dish: dishQuery,
-      totalCookingTime: cookingTimeQuery,
-    },
-    take: 20,
-  })
-  console.log(searchInfo)
-  console.log(recipes)
-  res.json(recipes)
-})
+app.post("/api/searchRecipes", SearchController.searchRecipes)
 
 app.post("/api/searchRecipes/keywords", async (req, res) => {
   const { keywords } = req.body
